# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления библиотекой",
        "title": "Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/loans": {
            "get": {
                "description": "Возвращает список всех завмов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Получение списка займов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loan"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает нового займа в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Создание нового займа",
                "parameters": [
                    {
                        "description": "Займ",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/loans/{id}": {
            "put": {
                "description": "Обновляет данные существующего займа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Обновление информации об займе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID займа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные займа",
                        "name": "loans",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет займ по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Удаление займа",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID займа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Author": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "BirthDate представляет дату рождения автора\nexample: 1828-09-09",
                    "type": "string"
                },
                "name": {
                    "description": "Name представляет имя автора\nexample: Лев Толстой",
                    "type": "string"
                },
                "nationality": {
                    "description": "Nationality представляет национальность автора\nexample: Русский",
                    "type": "string"
                }
            }
        },
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Author представляет автора книги",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Author"
                        }
                    ]
                },
                "author_id": {
                    "description": "AuthorID представляет ID автора книги\nexample: 1",
                    "type": "integer"
                },
                "available": {
                    "description": "Available указывает, доступна ли книга\nexample: true",
                    "type": "boolean"
                },
                "isbn": {
                    "description": "ISBN представляет международный стандартный книжный номер\nexample: 978-3-16-148410-0",
                    "type": "string"
                },
                "published_date": {
                    "description": "PublishedDate представляет дату публикации книги\nexample: 1869-01-01",
                    "type": "string"
                },
                "title": {
                    "description": "Title представляет название книги\nexample: Война и мир",
                    "type": "string"
                }
            }
        },
        "models.Loan": {
            "type": "object",
            "properties": {
                "book": {
                    "description": "Book представляет информацию о книге",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Book"
                        }
                    ]
                },
                "book_id": {
                    "description": "BookID представляет ID книги, которую берут в заем\nexample: 123",
                    "type": "integer"
                },
                "loan_date": {
                    "description": "LoanDate представляет дату выдачи книги\nexample: 2024-04-01",
                    "type": "string"
                },
                "reader": {
                    "description": "Reader представляет информацию о читателе",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Reader"
                        }
                    ]
                },
                "reader_id": {
                    "description": "ReaderID представляет ID читателя, берущего книгу\nexample: 456",
                    "type": "integer"
                },
                "return_date": {
                    "description": "ReturnDate представляет дату возврата книги\nexample: 2024-04-15",
                    "type": "string"
                }
            }
        },
        "models.Reader": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email представляет электронную почту читателя\nexample: ivan.ivanov@example.com",
                    "type": "string"
                },
                "membership_date": {
                    "description": "MembershipDate представляет дату вступления читателя в библиотеку\nexample: 2020-01-15",
                    "type": "string"
                },
                "name": {
                    "description": "Name представляет имя читателя\nexample: Иван Иванов",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}